split out feature profile CLI microcrate
