+++ crates/perl-lsp-feature-cli/README.md
@@ -1,0 +1,3 @@
+# perl-lsp-feature-cli
+
+CLI-facing helpers for parsing and validating Perl LSP feature profile arguments.
