+++ crates/perl-lsp-feature-governance/Cargo.toml
@@ -15,6 +15,7 @@ categories = ["development-tools", "development-tools::testing"]
 include = ["src/**", "Cargo.toml", "README.md", "LICENSE*"]
 
 [dependencies]
+perl-lsp-feature-cli = { workspace = true }
 perl-lsp-feature-contracts = { workspace = true }
 perl-lsp-feature-grid = { workspace = true }
 perl-lsp-feature-policy = { workspace = true }
@@ -23,6 +24,7 @@ perl-lsp-feature-profile = { workspace = true }
 [features]
 lsp-ga-lock = [
     "perl-lsp-feature-contracts/lsp-ga-lock",
+    "perl-lsp-feature-cli/lsp-ga-lock",
     "perl-lsp-feature-grid/lsp-ga-lock",
     "perl-lsp-feature-policy/lsp-ga-lock",
     "perl-lsp-feature-profile/lsp-ga-lock",
